#64 made Admin Controller and new routing for the admin
