+++ resources/views/volunteer/show.blade.php
@@ -63,8 +63,7 @@
         </ul>
         {{--check if the logged in user is admin to show the validation form --}}
             @if( Auth::user()->role == 8)
-                    @include('volunteer/adminPanel/validate')
+                    @include('volunteer.adminPanel.Assign Validator')
             @endif
-    </ul>
 
 @endsection
